Revert "Merge branch 'main' of https://github.com/Reself-C/Dataexpand into main"
This reverts commit e1e8839fa7e92df304132654f3e18029f535d7ee, reversing
changes made to 8395bc5656ff31b7acc4d20e2772f2e5e6a43898.

+++ DataGenerators/camera_utils.py
@@ -10,7 +10,7 @@ phone = {
     'tracking': True,
     'inmat': np.array([[1527.4,     0,  957.1],
                        [     0,1529.2,  529.8],
-                       [     0,     0,      1]],dtype=np.float32)
+                       [     0,     0,      1]])
 }
 
 others = {
